changed the adc period to increase buffer samples but maintain 100Hz frequency, converted 12bit adc reading to voltage and subsequently pressure in millibar

+++ sw/airborne/modules/sensors/turbulence_adc.h
@@ -30,7 +30,7 @@
 
 struct TurbulenceAdc {
   uint16_t raw;
-  uint16_t offset;
+  float offset;
   float scaled;
 };
